+++ docs/changes.rst
@@ -12,6 +12,7 @@ Released on March 26th 2024.
 * Superuser access to components with :ref:`component-restricted`.
 * Adjusted default :setting:`LOGIN_REQUIRED_URLS_EXCEPTIONS` to not block :ref:`manage-appearance`.
 * Avoid crash on pushing changes to diverged repository.
+* Avoid crash when installing :ref:`addon-weblate.generate.pseudolocale`.
 
 **Compatibility**